fix(analysis): use exact primitive type facts

+++ crates/vela_analysis/tests/completion_fixtures.rs
@@ -81,11 +81,11 @@ fn registry_facts() -> RegistryFacts {
     let mut registry = TypeRegistry::new();
     registry.register(
         TypeDesc::new(TypeKey::new(TypeId::new(1), "Player"))
-            .field(FieldDesc::new(FieldId::new(1), "level").type_hint("int"))
+            .field(FieldDesc::new(FieldId::new(1), "level").type_hint("i64"))
             .field(FieldDesc::new(FieldId::new(2), "inventory").type_hint("map"))
             .method(
                 MethodDesc::new(HostMethodId::new(1), "grant_exp")
-                    .param(MethodParamDesc::new("amount").type_hint("int"))
+                    .param(MethodParamDesc::new("amount").type_hint("i64"))
                     .return_type("bool"),
             ),
     );
